HW1/requirements.txt
# OpenAI Tools Dependencies
# Alternative to pyproject.toml for pip users

python-dotenv>=0.9.9
openai>=1.101.0
requests>=2.32.5
tavily-python>=0.7.11
yfinance>=0.2.65
